FDP-2758 - More flexible SignableMessageWrapper using delegation and backwards compatibility
